requirements.txt
fastapi[all]
python-dotenv
openai
requests
numpy
tiktoken
torch
nltk
transformers
pydantic